Arduino Direct Port Multistrober
This code is used to independently strobe/pulse two separate pins (2 and 4) at frequencies less than 1 millisecond, which is too fast for basic digitalwrite codes.

Code utilises direct port manipulation, and is based on the following tutorial
- https://forum.arduino.cc/index.php?topic=355427.0

In addition, this code allows for two strobes to concurrently run at the same time.
To achive this, two millis counters are recorded in two local states. And those states are used by 4 statements to read 1 of 4 potential global states
(ON-ON, ON-OFF, OFF-ON, OFF-OFF)